backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
pytest==8.3.4
httpx==0.28.1
python-multipart==0.0.20
pdfplumber==0.11.6
python-docx==1.1.2
lxml==5.3.0